check that no servers are running after every test
